Remove x11rb, move to platform implementations etc
