+++ Cargo.toml
@@ -121,7 +121,6 @@ gdk-pixbuf = "0.16"
 libappindicator = { version = "0.8.0", optional = true }
 dirs-next = { version = "2.0.0", optional = true }
 x11-dl = "2.21"
-x11rb = { version = "0.12.0",  features = ["allow-unsafe-code", "dl-libxcb"] }
 dbus = "0.9"
 uuid = { version = "1.3", features = [ "v4" ] }
 png = "0.17"
