+++ src/platform/unix.rs
@@ -21,6 +21,7 @@ use crate::{
   event_loop::{EventLoop, EventLoopWindowTarget},
   platform_impl::{x11::xdisplay::XError, Parent},
   window::{Window, WindowBuilder},
+  TaskbarProgressState
 };
 
 use self::x11::xdisplay::XConnection;
@@ -32,6 +33,16 @@ pub trait WindowExtUnix {
 
   /// Whether to show the window icon in the taskbar or not.
   fn set_skip_taskbar(&self, skip: bool);
+
+  /// Sets the taskbar progress state./ 
+  /// ## NOTE
+  /// - On **Linux:** App must be installed using .deb binary, Might not work on some distros like Linux Mint (Cinnamon)
+  fn set_taskbar_progress(&self, current: u64, total: u64, unity_uri: Option<&str>);
+
+  /// Sets the taskbar progress state./ 
+  /// ## NOTE
+  /// - On **Linux:** App must be installed using .deb binary, Might not work on some distros like Linux Mint (Cinnamon)
+  fn set_taskbar_progress_state(&self, state: TaskbarProgressState, unity_uri: Option<&str>);
 }
 
 impl WindowExtUnix for Window {
@@ -41,6 +52,30 @@ impl WindowExtUnix for Window {
 
   fn set_skip_taskbar(&self, skip: bool) {
     self.window.set_skip_taskbar(skip);
+  }
+
+  fn set_taskbar_progress(&self, current: u64, total: u64, unity_uri: Option<&str>) {
+    let uri = match unity_uri {
+      Some(a) => Some(a.to_string()),
+      _ => None,
+    };
+    self.window.set_taskbar_progress(current, total, uri);
+  }
+
+  fn set_taskbar_progress_state(&self, state: TaskbarProgressState, unity_uri: Option<&str>) {
+    let uri = match unity_uri {
+      Some(a) => Some(a.to_string()),
+      _ => None,
+    };
+
+    let taskbar_state = {
+      match state {
+        TaskbarProgressState::None => false,
+        _ => true,
+      }
+    };
+
+    self.window.set_taskbar_progress_state(taskbar_state, uri);
   }
 }
